[rust] Do "cargo check" for library targets
Doing a "cargo build" on a library target generates a .rlib artifact,
but that's not actually used when a library is a dependency from
another target; cargo rebuilds from source.

This change makes library targets just do "cargo check", which is a
lot faster, especially for release builds.

TO-233 #done

Change-Id: I36916cf712eeb8ab34a1cdba0822b319b0c620fd