[ddk][iotxn] remove phys_offset, add iotxn_clone_partial(), clarify docs
Remove the phys_offset field. iotxn->phys[0] will point to the the
closest page to vmo_offset, rounded down. iotxn->virt will point
to vmo_offset.

Add a new function, iotxn_clone_partial(), which will clone a iotxn
but update the cloned iotxn's vmo_offset and length. The new
vmo_offset must be >= the original iotxn's vmo_offset. The new length
must be <= the original iotxn's length. This is useful for a
midlayer driver to clone an iotxn and restrict the data it operates
on, or for it to split a large data transaction into smaller pieces.
iotxn_clone_partial() updates the iotxn->virt, iotxn->phys and
iotxn->phys_count fields if they are set in the original iotxn.

iotxn_clone() and iotxn_clone_partial() will clone into the iotxn
pointed to by the out parameter if it is nonnull, instead of always
allocating a new one or obtaining one from the free list.

Add some comments in iotxn.h to clarify who is allowed to modify
each field in the iotxn struct.

Change-Id: I42b923a38de8e5514c3666b38158d46aa4a682e0