Use maxwell::ComponentScope when request IntelligenceServices.
The maxwell::ComponentScope is a FIDL struct that defines metadata about
the "access scope" for services requested by clients through the
UserIntelligenceProvider.

It should eventually be replaced with the Fuchsia-wide primitives that
satisfy the same role. For now, it is slightly more flexible than always
passing N optional strings around to define if a service is meant to be
for a Module in a story, or an independent Agent, etc.

Change-Id: I747710de30075bf401ba056bb4d480af7603dc04