[ddk][iotxn] Fix test failure and problem with iotxn_alloc()
Test was failing because we were not initializing out parameter to NULL
before calling iotxn_clone().
In this case, the clone ended up being the same as the txn being cloned,
resulting in a hang due to free list corruption.

Also fixed iotxn_alloc() so it will set release_cb properly in the case
where it pulls a txn out of the free list.

Change-Id: If3a188dbffff6419e543909e51caafb4cfe1b51d