[tftp] Refactor sending to allow for out-of-order ACKs
If the receiving side sends an ACK other than for the end of a window,
the sender should rewind and start a new window based on this ACK. The
previous API did not allow for non-blocking reads in between sends, so
the library no longer attempts to send data.

read and write callbacks are used instead of memcpy and a send callback,
allowing the client of the library to use the tftp data however they
want (e.g., file, vmo, etc).

Timeouts still aren't handled correctly, but require some more
refactoring of the state machine.

Change-Id: Ie98bdf8d7761ca7835691695ef840ffd9947c4bc