[fs][memfs][minfs] Allow readdir to operate on modified directories
This patch fulfills a well-known requirement of POSIX, that readdir
will return all entries which have not been (1) added or (2) removed
since the last call to "opendir" or "rewinddir". Previously, if
readdir was called on a large enough directory, it would simply use
a "sequence count" on the parent directory to track the boolean
concept "Has this directory been modified since our last call to
readdir?" If yes, readdir would return ERR_IO rather than returning
all unmodified entries.

This patch provides a notion of "ordering" to entries in a directory
for both MinFS and MemFS, so they may preserve an absolute notion
of this ordering between calls to readdir.

- If the ordering of an entry in the directory is less than the readdir
cookie's order, then (if it has been unmodified) it has already been
returned.
- If the ordering of an entry in the directory is greater than or equal
to the readdir cookie's order, then it should be returned, since
it has not yet been seen by this call to readdir.

MinFS and MemFS use different techniques to describe "order". Since
MinFS uses a flat directory layout, "order" is simply the offset
into the directory (as it was with the original Unix filesystems when
readdir semantics were originally defined). MemFS, however, stores
its entries in a linked list, providing a non-linear directory.
To define an "absolute order" among MemFS entries, when nodes are added
to a directory, they:
- Are appended to the end of the direntry linked list
- Hold a token which is always greater than the previous direntry

US-214 #done

Change-Id: I4aafbad904d2ac4bd4b7eb18815adcbbaa3930af