[gap] Introduce Adapter class
This CL introduces the Adapter class which represents the host+controller
state for a HCI device. This class is responsible for putting all the
various subsystems together and is intended to be the root entry point
for all GAP operations.

For now this only performs some parts of the initial Low Energy bootstrap
sequence.

Change-Id: I6f57fade50766fa1de08d0c1c495adf9ac3f8c1f