Assemble sysroot in GN build
This adds GN rules that assemble a sysroot containing the artifacts from
the Magenta build's exported sysroot and the C++ standard library
artifacts. These rules are configured as toolchain dependencies of the
Fuchsia toolchain so they will run if necessary whenever the inputs,
either Magenta or C++ standard library, change. This ensures that if the
Magenta build is up to date all of the Fuchsia build will match,
eliminating one class of mismatch. Additionally since ninja is tracking
the actual file changes in the sysroot it can correctly perform
incremental rebuilds as needed.  There are a few pieces at play:

The C++ standard library is provided as part of the toolchain prebuilts
downloaded by a jiri hook and places into //buildtools/toolchain/...
The jiri hook updates a stamp file as well. The sysroot assembly step
has a script that enumerates the files in the sysroot to figure out
which files to copy into the assembled sysroot and where to put them.
This enumeration step declares an exec_script dependency on the stamp
file, so it runs (along with the rest of GN generation) whenever a new
toolchain is downloaded.

The Magenta sysroot artifacts are provided in a similar way, although
instead of them being downloaded by a jiri hook they are built by an
earlier build step into a well-known location that also exposes a
file listing the contents of the sysroot.

Change-Id: Ibde356a6377c739d485cc787fd3ea4d5161ec174