[kernel][magenta] kernel policy refresh
Kernel policy refers to the things that are not
covered by object-capabilities (aka handles). For example
how to restrict channel creation or how to control
memory usage.

We had an exploration of how that would look in the
"bad handle" policy, now its time to get a real version
this time:

- Centered in the job object (only way to set policy)
- Inmutable: the policy applies when processes are created
  and it will never change.

Actually in the CL:

1- remove the "bad handle" policy as it was implemented
2- introduce the new policy visible parts which form
   a set of key-value pairs [topic:policy]
3- introduce two new rights set-policy and get-policy
4- a fake test to drive the point home

Change-Id: I2858ca615d6b9cefdd0896243f915a1c527d0187