[ddk][iotxn] Add utility for iterating through physically contiguous ranges
iotxn_phys_iter_t is a utility to make it easier for drivers
to implement scatter/gather support.

iotxn_phys_iter_init() initializes a iotxn_phys_iter_t for iterating through
the physically contiguous ranges in the iotxn's buffer

iotxn_phys_iter_next() returns the next physical address and length of
the next physically contiguous range, or returns false if we have reached the
end of the buffer.
iotxn_phys_iter_next() takes a "max_length" parameter, which can be used to
limit the size of the range returned to whatever limit might be required
by the driver. max_length must be a multiple of PAGE_SIZE.

Tests added to iotxn-test.c

Change-Id: Iec0d5ef8d26e472e03c15e7509fd421a5d692014