Set rpath to $ORIGIN on linux so host binaries can find dynamic libraries we build
We compile and run several helper and generation programs on the host
during the build. Some of these programs depend on libraries that we
want to deploy as shared libraries on the target (i.e. zlib). To keep
the build specification simple, we configure these libraries as shared
libraries in all configurations even though it only really matters for
the target. Thus, host binaries running on linux need to be able to
find the shared libraries next to the binaries in the out directory.

This approach doesn't work as well for any host binaries we may wish to
distribute in the future. When we need that, we'll have to think more
carefully about how to link and distribute the host bits.

Change-Id: I55b9cfcd2a3c8f83e1d71d7471b0b4e0b5f49caf