[gfxconsole] Simplify scroll-to-bottom-on-keypress behaviour
The previous behaviour was complicated: If the user presses a key, the
console will scroll to the bottom the next time there's output on the
console.  This applies even for modifier keys.  Scroll-to-bottom stops
if there's unconsumed input in the input buffer.  (So if the program
on the console isn't reading input, scroll-to-bottom works for the
first non-modifier keypress but stops after that.)

Simplify as follows: Any time there's input (which excludes modifier
keypresses), immediately scroll to the bottom.  This matches
gnome-terminal.

Change-Id: If02ad82bbdc2419f5f433e72dc72bf3689b46161