[gfxconsole] Fix implementation of circular scrollback buffer
As lines are added to the scrollback buffer, we'd expect the result of
vc_device_get_scrollback_lines() to increase until it hits 1024 (the
scrollback limit), and then stay at 1024.  Instead, before this
change, it would drop down to 0 and start increasing again.

Fix this by reimplementing the circular buffer logic.  Using head and
tail indexes has the problem of whether head == tail means the buffer
is full or empty.  Instead, use head and size values.

MG-681

Change-Id: I36bf2474b888e9d3612f9e81a020043e4b2444e7