[devmgr] phase 4: create() driver hook for bus manager devices
The create() hook will be used to ask a bus manager driver to
instantiate a shadow instance of a bus device in a new devhost.

Build pci bus driver standalone in a way that works with the
phase 4 world.  The phase 3 world continues to use the built-in
version.

Fix a bug in message unpacking.

Change-Id: I63e0338a0071c31b242eb69ac434e12aab053f0b