[utest] run magenta tests in groups
Assign a group to magenta tests, and run tests in order of
increasing complexity, and abort the test if a group failed.
This is because if there is a bug in a core magenta facility,
it could cause all higher level tests to fail.

The current groups are:

* "core": tests on kernel facilities
* "ddk": ddk tests
* "sys": general magenta system tests (default)
* "fs": filesystem tests

Set the group in the tests' rules.mk by setting the
MODULE_USERTEST_GROUP variable.

runtests accepts a list of groups to run. If none is specified,
run all tests.

Change-Id: Ia363d25f5c1c828a3d1a9abf586170bb90c679f7