app: allow the framework to control sync settings per-user
This patch allows the device runner to manage and specify the Firebase
server ID to be used for syncing the user data. For the transitional
period, we still default to the global config file if the device runner
passes an empty server id. (null means "disable sync")

Once the neccessary plumbing is added in the device runner and we
eliminate all remaining uses of the config file, we will drop it
altogether.

LE-183 #comment allow the framework to control sync settings per-user

Change-Id: Ifaee3de2dedbe46c8027d84dcacacf6157f8eeb0