Explicitly trigger rebuilds of all targets built for fuchsia when toolchain updates
This adds a define line with the toolchain version (from the stamp file
written out by jiri) to the compilation line for all targets built for
Fuchsia, so we always rebuild when the compiler or other parts of the
toolchain update. Today this is (mostly) redundant with the sysroot
stamp file, but since that is going away soon we need a mechanism to
trigger rebuilds for compiler changes that don't change headers or
libraries in the sysroot proper.

Change-Id: Icf9798002e7537b74dadc1bb9dff4da662e2201d