[kernel][magenta] flush out policy design
This CL has the basics of the job/process policy engine.

The basic concept is that while we might have many jobs
and processes, the number of unique policies is low, like
say in the order of 20.

So rather than burden each process with arrays of
mx_policy_basic_t we have a PolicyManager that achieves
the following:

1- simple policies are encoded in the policy handle (cookie)
2- oddball policies are centrally mantained and the handle
   points to it.

This also makes creating child jobs and processes
fairly fast, or better said, the slowdown should be minimal

Change-Id: I83b27c2b82fe7e2d60fa41439ea51d17bb369159