[hci] Fix potential leaked references in CommandChannel callbacks
CommandChannel was not immediately clearing the transaction data
(completion and status callbacks) of a pending HCI command at the end
of the transaction, deferring the clean-up until overridden by the
next HCI command.

This caused a potential leak of data attached to the callbacks if
a CommandChannel got used to send a single HCI command. This could
lead to a circular reference if a ref-counted object that owns the
Transport (and thus the CommandChannel) attaches itself to a HCI
callback, expecting the callback to relinquish the reference as soon
as the transaction is over.

Change-Id: I82307fbc2da57ebf5a355495825329f53e5f7cbb