[fs] Add RemoteContainer, making the 'remote_' field optional
Previously, fs::Vnode has a 'remote_' field, which could
be used to mount remote devices, services, and filesystems on
top of parent nodes. For MemFS, MinFS, and other traditional
filesystems, this provided an easy mechanism to get the
ability to mount sub-filesystems for free. However, for
other filesystems (service FS, blobstore), mounting is not
a desirable functionality, and the "remote_" field is wasted.

This patch creates the "RemoteContainer" class, which can be
used by Vnodes that want to host remote filesystems.
Additionally, this patch updates MinFS and MemFS directories
to use this RemoteContainer.

Change-Id: Ic56e84a82f01503119f16e39c70967b14cb9cbc5