[gfxconsole] Ensure that output gets displayed while the viewport is scrolled
Fix vc_device_write() to adjust for the viewport scrolling offset.
The invy0/y1 range is relative to the top of the main (non-scrollback)
console area, but vc_gfx_invalidate() takes row numbers relative to
the top of the screen, so vc_device_write() needs to adjust for that.

MG-681

Change-Id: I8b7622f3d81b4771eed5597374bfe3222e6d52c8