apps/netconnector: add MdnsAgent and agent host
This CL adds MdnsAgent, an abstract class that defines a component
that participates in mDNS question/resource traffic. It also provides
an implementation for the 'host' for agents, the Mdns class.

MdnsAgent implementations see all incoming questions and resources,
can send questions and resources and can asked to be awoken at
specific times. The host owns the MdnsTransceiver and hosts a
collection of agents, routing question and resource records and
managing task execution.

Future CLs will add MdnsAgent implementations that do things like:
- respond to address queries
- publish a service instance
- subscribe to a service
- probe for name conflicts
- renew expiring records

Change-Id: Id30a8414590f9e67e5f74c5990b9005d4c0830f5