[gfxconsole] Fix scrolling while the viewport is scrolled
Fix the update of the display if console output causes the display to
scroll while the viewport is scrolled.

Originally, vc-device.cpp had various checks that do "if
(dev->viewport_y < 0) return;", which attempt to stop console output
from being reflected on the screen while the viewport is scrolled.
However, that logic was flawed because some output would get displayed
anyway, and the display would get left inconsistent.  This change is
part of fixing the display update inconsistencies.

MG-681

Change-Id: Ie0770cfe3adf730d75b2dec4173519b461e8b189