[gfxconsole] Fix non-cursor-moving display updates when viewport is scrolled
Originally, vc-device.cpp had various checks that do "if
(dev->viewport_y < 0) return;", which attempt to stop console output
from being reflected on the screen while the viewport is scrolled.

vc_tc_invalidate() has such a check, but it's mostly ineffective,
because most console output moves the cursor, which also has the
effect of updating the display.  This check is unnecessary, so we can
remove it.

We can test this, because some control sequences (those for erasing
characters) change the console contents without moving the cursor.
This change therefore fixes a bug in that case.

MG-681

Change-Id: I298eabd417f24910c8311386288e12e43ffb50e0