[gfxconsole] Fix display update consistency at the scrollback limit
Before this change, if we're displaying the top of the scrollback
buffer when lines are dropped from the scrollback buffer, those lines
are left on the display, but scrolling the viewport down (with
Shift+PageDown or Alt+Down) can produce an inconsistent display.

Fix this by triggering a display update: Stop displaying lines that
were dropped from the top of the scrollback buffer.

MG-681

Change-Id: I3e14a0ff36fe1dcb35b5ba8e934be1d76bedecf5