[xhci] Add a debug printf.
Add a debug printf which fires when we see a condition code in a
transfer event that we do not have a handler for.  The stack is going
to close the connection to the peer and it is helpful (when debug
printfs are enabled is turned on) to know why the stack decided to
close the connection.

Change-Id: Ib7ef03acc7bc256478a2c416a6c39291bb9c1dd5