[blob] Combine Blob and Vnode objects
During the move to RefPtrs, this turned out to be a
problem, as Blob objects point to VnodeBlobs and vice
versa, causing a circular dependency.

Since the lifetimes of these objects are so tightly linked,
this patch combines them into a single object. It still
keeps "VFS"-like operations in "blobstore-ops", with the
internal versions in "blobstore.cpp".

MG-719 #comment In Progress

Change-Id: I73f20c2de970d13e8c52c1e5d4a125fa70ae3496