[udev][usb-hub] Add support for USB 3.0 hub features
USB 3.0 adds a number of new feature selectors. If we don't send a
CLEAR_FEATURE command to acknowledge when they change, the hub floods
us with a continuous of stream of interrupt events to keep notifying
us of a port's status. In particular, in one case we keep receiving
events with the USB_C_BH_PORT_RESET bit set. We now acknowlege this
event by sending CLEAR_FEATURE(USB_FEATURE_BH_PORT_RESET) to stop the
stream of events.

Also, in USB 3.0 the wPortStatus bits no longer correspond exactly to
the wPortChange bits, so we add separate defines in hw/usb.h for them now.
For example we now have USB_PORT_CONNECTION bit for wPortStatus and
USB_C_PORT_CONNECTION for wPortChange.

Change-Id: Ib7de368f9063b5979f844a59655f9e25faab13e6