[kernel] Generate --build-id note in magenta.elf, use in print_version
This changes the kernel link to use --build-id, with symbols
that make it easy to find the note from inside the kernel.
It also consolidates some linker script cleanups.

A hex-formatted string of the build ID bits is added to the
version struct that's printed out at boot in debug builds.

Change-Id: I24be44b8599be956af578d29edffe3a25f002bfb