Have default fuchsia configs be configurable
It turns out if a config includes another config via its 'configs'
variable, those nested configs can't be removed later on - only configs
in the top-level config list or those set via set_defaults() can be
toggled. We have a number of configs we want to be set by default for
Fuchsia targets but may need to be disabled by various parts of code.
So this is possible, this extracts those from the
//build/config/fuchsia:compiler config up to the
_shared_binary_target_configs list which is applied to the default set
for source_set, etc, through set_defaults().

Change-Id: I4ab1db297796632d4a615295e3aa7d40481e59a1