[hypervisor] Handle TSC deadline timer in guests.
When the TSC deadline MSR is modified, we save it to the local APIC
state. We also instruct VMX to VM exit on the HLT instruction. On HLT,
we set the thread to sleep until the deadline passes. Once the
deadline has passed, we inject an APIC timer external interrupt into
the guest and continue.

I should validate that the APIC timer was actually set, but I need to
investigate why the virtual APIC page isn't reflecting local APIC
writes.

Change-Id: I9d8277fc2e886ffaeda9ec39578977ed7d28532f