[fs] Test ThinFS whenever it is available
It's a little tricky to test ThinFS; it won't exist on a
pure Magenta build, though it will exist on a variety
of Fuchsia builds. Update the fs tests to run the test suite
on ThinFS whenever it is detected in "/system/bin/thinfs".

Additionally, update the "attr" test to adjust the expected
timestamp to the underlying filesystem granularity.

Change-Id: I949c4c0e0a3fe04abaf8bf1dbd541a5f34812c0f