apps/netconnector: add GetKnownDeviceNames and --show-devices
This CL adds a method to the NetConnector interface that allows
clients to get an array of known device names with the option of
getting updates with a long 'Get'. It also adds the --show-devices
option to NetConnector (when running as a utility) so the known
devices can be obtained from the command line.

Change-Id: I4b9a5357fa7bf3c595f9ecf174c94eaeafd811e2