[common][hci] Refactor DeviceAddress
This CL breaks the existing DeviceAddress class into the
DeviceAddressBytes and the new DeviceAddress class. While the former
continues to serve as a serializable 6-octet address structure, the
new DeviceAddress also incorporates an address-type field.

DeviceAddress can now also be hashed and used as a key in ordered
and unordered STL associative containers.

Change-Id: I50e62c78dde034a42083a6673b3a0081315a23e5