Fix crash in LinkImpl.
When a link becomes orphaned, it cannot be deleted before previous
updates to it have been written to the ledger, and change listeners
be notified, because the notify call accesses the LinkImpl instance.

Thus, we invoke the orphaned listener in a callback of Sync(). Also
we check for no connections in the callback of Sync rather than before
Sync, because new connection may appear now that we always connect
to links by name (this was wrong already before the recent change
to referencing link by name rather than by handle, but now it's more
likely that a link gets orphaned).

The fix here is still not correct in the case that a new connection
is made after Sync() is invoked, is written to, and disconnected. It's
unlikley but not impossible. Left a TODO().

Also, when a link impl gets deleted, we must remove the on change
watchers from the story storage impl.

FW-169 #done

Change-Id: Ia710c0ffc75f5a0ca3e4592970f711276eb2b125