Fix for MG-742
Fix thread annontations for the pending and in-flight CORB queues.
Once those have been corrected, add protection for the CORB in-flight
queue during RIRB processing.  This prevents the IRQ thread from
de-queueing queued CORB commands when processing the response ring
buffer at the same time that a client might be queueing a command to
the CORB.

MG-742 #done

Change-Id: I63fbcbc4ccea9e0b37460a2252fbfeff13cf8b68