[udev][block] normalize error handling
read() and write() on block devices perform truncated transactions
if the requested length is unaligned or if it attempts to access
beyond the capacity of the device. BLOCK_CORE operations do not
truncate and return ERR_OUT_OF_RANGE for out of range accesses,
and ERR_INVALID_ARGS for unaligned accesses. Unaligned device
offsets always return ERR_INVALID_ARGS.

Change-Id: Ia0b99e0d37f1caed77c267f4efda532ba0a5a6ca