Dart cleanup
Making the Dart analyzer happy. Also, adding to the README about the context
dashboard.

Much of the analysis failure seemed to be due to a change in the Fuchsia Dart
analyzer plugin to only consider BUILD.gn's in the immediate Dart directories,
whereas some of our Dart directories were goverened by BUILD.gn's in parent
directories.

Change-Id: I3bf02035ea7e10af70d582c2eb4411d64d46f8a1