[scripts][run-magenta] Use different mac address for different interfaces.
By default, Qemu always use the same mac address for a given interface.
If Qemu is run multiple time with different bridged tun interfaces, all
the interface will have the same mac address.

This changes the run-magenta script so that the mac address is deduced
from the name of the network interface.

Change-Id: I3e4162228e6b400b5b9d219545201258f5755bc9