[sysroot] Cut down on extraneous rebuilds after Magenta changes
This avoids rebuilding targets in Fuchsia when Magenta internals
change. There are two parts:

*) Avoid touching the assemble_sysroot stamp unless we actually changed
something. Since this rule has restat = 1, if we do not touch the stamp
ninja will notice that we did nothing and not re-run steps that depend
on this step.

*) Avoid copying debug-info from Magenta into the sysroot as this isn't
needed in order to compile or link anything in Fuchsia and changes
whenever Magenta's internals change.

Change-Id: I433919285a9fb6fcf1814ec1113dcd7aced17a0c