Rename build-sysroot.sh to build-magenta.sh, update wrappers
This renames build-sysroot.sh to build-magenta.sh since the only thing
it does is build magenta into the location Fuchsia expects, and makes
it quiet by default and closer to idempotent as well.

This is almost fast and smooth enough to run by default on every single
build - on my MacBook Pro the up-to-date build still takes 2.7 seconds,
which is a bit too slow for my taste. I'll add a step in front to
short-circuit the build when no files as changed as a follow up.

This also deprecates the fbuild-sysroot{,-if-changed} wrappers now that
they are identical to mbuild.

Change-Id: Ie69a70779f7e460e285ebd9efcb0fcdbd7274b96