Explicitly fill the trailing portion of directories with zero
If directories are not filled with zeroes after the "last entry"
marker, then reasonable FAT filesystems should treat this
area as "undefined". However, fsck actually checks that the
trailing end of directories follows the specification word-for-word,
and as a consequence, it may complain if any out-of-bounds
(but still intentionally freed) direntries are found.

Change-Id: Id14daa2472f09dcd36241351113e29f68880d8a8