Add C++ runtime libraries as a package configuration instead of hardcoded
This moves information about the source and desired build location from
being hardcoded in process_modules.py to being in a package
configuration file. This configuration is hardcoded as added to the set
of configurations to process since approximately every project depends
on it now, but we can stop doing this when we teach the system more
about runtime dependencies.

Change-Id: I1d09e7234e0fbffaaeeba20f2ffa05637ab29636