[minfs] Simplify Inode destruction
Remove an unnecesary metadata write to the inode table
during inode destruction. If the bitmap identifying
available inodes indicates that an inode entry is not in use,
the contents will be ignored. Once that inode is allocated,
it is explicit set to zero anyway (see: InoNew). Thus,
writing back zeroes to the inode table simply amplifies
writes.

Change-Id: Iaee51329d39b289d363294850e8e6acd1b3bbceb