Teach installer about /boot manifest from Fuchsia build
In order to temporarily work around a device manager load order issue,
the Fuchsia build produces several files that need to be loaded in
the /boot partition on the installed system. Since the installer parses
the manifest file from the Fuchsia build, it needs to know about the
boot manifest file as well and append it to the bootdata it installs to
the EFI partition.

Change-Id: I6ea452b1c0c83eb8ba367072ade42af091343f5d