apps/media: improve example app command line processing
This CL changes the command lines for audio_player and media_player
example apps so that URLs or file paths are supplied positionally.
Previously, a URL had to be supplied (file:///... for files) using
the "--url" option. The new regime takes advantage of automatic
completion in the shell when file paths are used.

The README.md for media_player_flutter has been updated to improve
clarity.

Change-Id: I11a9eff1db736b5926f041014d0cee47e61b41a7