[context] Add focused Story aliasing.
This adds a Coprocessor that tracks /story/focused_id and duplicates
the context in /story/id/<focused_id> to /story/focused.

For now it's just a copy, but we may want to build it into the
ContextRepository such that we're not copying data. However, it'll be
fine until we can show it hurts us.

Change-Id: Id19a9add1c19d0e1d2d42186a52035aefc1a5514