More test life cycle fixes.
Extract elements of life cycle management of a component in a test
into a common base class ComponentBase<ComponentType>, where
ComponentType may be Module, Agent, or UserShell.

This base class contains a weak pointer factory that guards timeouts
posted to the message queue, and the method DeleteAndQuit() which
deletes the instance, posts testing::Done() to the test runner and
then quits the application.

All test agents and modules now use this base class to implement their
life cycle. Because they now all delete themselves before sending
testing::Done(), all their test points are indeed reported, which
was not the case previously.

This still does not fix the triggers test for me, which continues
to fail in my qemu.

Collateral cleanup:

* Mark one test agent testonly.

* Drop empty parens from constructors invoked with new.

* Make base classes private where their interface is not exposed.

Change-Id: I226ca7da211f8e465ebcb39f25e864e5416b48c8