apps/media: handle imprecise audio PTSs
This CL addresses a bug each in the audio and video decoders and also
compensates for audio PTS values that aren't sufficiently precise.

The bug in the video decoder is that it didn't handle inconsistent
incoming pts rates. This never manifested as a problem, because
ffmpeg doesn't change pts rates. In theory, it should be supported.

The bug in the audio decoder is that it assumed that incoming packets
use a 'frame rate' pts rate e.g. 48000/1 for 48k audio. This was
causing audio rendering to fail with some content.

The other issue is that we're seeing audio packets with a pts rate
of 1000/1, meaning that the pts's are only in ms resolution. This
isn't adequate precision to actually reflect when the packet should
be played out, so we have to synthesize our own PTSs.

Change-Id: I430bcceda33f3079d6f3d3bda1fe754dfe8e48ca