Permit multiple concurrent stream client connections.
Allow multiple client to connect to a stream concurrently.  It is not
clear if we are going to allow this or not in the long run, but until
we have an audio policy manager, we need a way to change the hardware
gain if the main mixer currently has the stream channel open.

Change-Id: Id46a8aeb76161ee53fb4d2a67c8317c9c5b1811f