[kernel][magenta] last handle signal
Introduces a new signal MX_SIGNAL_LAST_HANDLE which
is on by default on new created objects.

Moves all the heavy logic of 0 handle, one handle
and so on to magenta.cpp which is already taking
the very convenient lock. This also removes
the atomic operations on the handle count, however
in order to properly synchronize the handle
count and the signals, a fairly ugly
double check is introduced in the state_tracker
via StateTracker::UpdateLastHandleSignal()

- Adds a new .md that better explains generic signals
- Adds new tests for the corner cases.

The rest of the code is fixing the many tests
that compare signals with a particular value which
need to be updated for the new signal that
shows up unless the handle is duplicated.

Change-Id: Id42b3fe951c16407bc08e18bb342ddd8f2830bf9