[env.sh] Unconditionally build Magenta in fbuild/etc
When the Magenta + sysroot building step always ran a number of slow
steps, it made some amount of sense to have heuristics in env.sh that
avoided running this step sometimes. Now that the sysroot build step is
folded into ninja and the magenta build used by this setup is fast for
no-op builds, this extra heuristic only gets in the way.  Running
build-magenta.sh with an up-to-date build takes ~1.5s to determine that
it has nothing to do on a fast Linux box and ~2.7s on a slow-ish Mac.
If we want to optimize this out of the normal fbuild cycle, we can do
this inside build-magenta.sh in a more precise way.

Change-Id: Iea4c62615215553bd49505de8d92519a02f3014c