Hide use of weak pointers in ComponentBase.
Instead of exposing GetWeakPtr() to derived classes of
ComponentBase, expose a method Protect() that
takes a callback and returns it wrapped into a check of the weak
pointer that prevents the callback from executing if it's invoked
after `this` is deleted.

TODO in followup:

No change was made to what the scheduled callbacks do, although
this could be unified in the future too. Some quit outright,
some just request Stop() from the owner. Both is viable, but the
variety should perhaps be covered more systematically. Also,
most is test timeout protection, which could be moved into
the harness altogether.

Change-Id: I3ac28f5da0370545d4d1dea4936f3ab9cf1d2758