[pcie] Expose device info from PCIe protocol
Rename the basic device info structure from mx_pci_get_nth_info_t to
mx_pci_device_into_t.  Expose this information via a new PCI device
protocol method; "get_device_info".

This information is always known to the bus driver because represents
the set of properties which need to be published to the device manager
for it to evaluate driver binding scripts.

Most of the information, however, is useful to device drivers as well.
Things like the hardware interface revision, the VID/DID, the device
class (and so on) get used by drivers to deal with hardware quirks,
understand hardware capabilites, etc...  Except for the BDF address of
the device, drivers can get this infomation via the device config, but
it is so common to want to access this info, that it seems like a
shame to force devices to map their config, know the structure of PCI
config, and read all of the elements on their own.  Instead, they can
now just fetch it via the PCI device protocol.

Change-Id: I2761ef27221ea80fe8bea5c419910d2ff106c785