Convert more Operations to using FlowToken.
Surfaces another subtlety: The destructor of the FlowToken may
actually be invoked *after* the Operation is destroyed:

* An Operation executes a fidl method on a connection that it
  does not own and puts the flow token on the capture list of
  the callback.

* While the callback for the method is outstanding, the owner
  of the operation queue gets destroyed. The pending Operation
  in the operation queue gets destroyed. The pending callback
  gets destroyed too and invokes Done() on the Operation, which
  however is already deleted. Chaos ensues. This often manifests
  itself by OperationQueue::Drop() failing with DCHECK.

This problem appears specifically if the method is invoked on an
InterfacePtr that is owned by the parent of the OperationQueue.
But it would also be a problem if the interface ptr is owned
by the Operation, because Done() would still be invoked after the
Operation destructor proper.

The solution is for the flow token to keep a weak ptr to the
Opeation, and gate Done() on the weak pointer being valid.

It's important to realize that this weak pointer does not
protect against access to Operation from fidl method callbacks
after the Operation instance is destroyed. This is already done
by design of Operation and the OperationQueue that suppresses
callback invocation after delete. However, it protects against
Operation being accessed from *destructors* of callback functions.
The destructor of the callback function (and thus of everything
on its capture list) are still invoked even though the callback
itself is not invoked.

Another newly encountered situation is when a flow of control needs
to continue along exactly one of multiple copies of the flow. In
that case there must not be multiple copies of the flow token,
because it calls Done() only when the last copy destructs. The
solution is then to put the token in a container that is shared
between the different branches. It acts essentially as the call
once flag. This is implemented in the new class FlowTokenHolder.

Discovered and worked around LE-213: Page.StartTransaction() is
not pipelined but always continued in the result callback. LE-213
was fixed before this CL was submitted.

Collateral cleanup:

* Some flows were missing Done().

* More uniform error log messages.

* const

* &

Change-Id: Ib638839547b151396cbac075333e09810cdea80c