apps/netconnector: allow local clients to access local services
Netconnector doesn't currently allow local clients to connect to
services that are provided locally. It really should allow this,
because that provides a consistent way for clients to get access
to running instances of services by name.

Local clients get local services using GetDeviceServiceProvider.
The device name can either be the name of the local devices or
"local".

Change-Id: I9434a6a263161eb389e5786484328d008992491c