Allow device shell to pick user shell.
This patch exposes an optional parameter in UserProvider.Login() which
is of type AppConfig. This AppConfig can be set by device shell to
anything it wants the framework to run as user shell. If not set, device
runner will pick a default based on command-line arguments.
This is not the full solution and it'd like to revisit this after
understanding requirements, especially around persistence

FW-180 #done

Change-Id: I03f8c8ed761f85808a89e8897062a2cab1dc4e9b