[minfs] Force fsck to return an error on bad state
Since fsck is being used in tests, to verify that
the filesystem state has not been corrupted, it is
important that any invalid filesystem state is reported
to the caller. To this end, whenever a warning is logged,
identify the filesystem state as "ERR_BAD_STATE",
and return the state if no other more significant
errors have occurred.

Additionally, update "CheckFile" to work with sparse
files, where the inode size may be larger than the
allocated number of blocks (multiplied by the block size).

Change-Id: Ie64de511f5b95e9a275c4e8ac16ca2996311663d