[kernel][portsv1] fix a bug that could sometimes cause waiters to get stuck
If a writer signalled the event too quickly in sequence, and it was at
the same time racing with a few waiters, there is a chance that an edge
would be lost and a waiter would get stuck.

Switch the event over to a non autounsignal and have it track the state
of whether or not there are any messages queued. Not as efficient, since
it'll cause waiters to wake up and stampede if more than one is
queued, but ports v2 is about to replace this anyway.

Change-Id: I3a236df342f135a81d2a8f835a7179455a4e716d