apps/media: improve renderer instrumentation
This CL enhances the instrumentation applied to renderers. Priming
and timeline changes are logged as well as packet 'engagement',
meaning the actual consumption of packets by the renderer. Packet
engagement is summarized in the digest in terms of how far in
advance of engagement packets arrive (earliness) and when they
arrive late (starvation).

Subsequent CLs will use these flog changes to instrument the audio
and video renderers.

Change-Id: I9202b8d7fc09d5a06ecf61eb1474dd0a0f446173