[gfxconsole] Fix type of argument to vc_do_root_open()
A recent change (133065be33ae1233f19fe50bdbab4a8adbf778dd) changed
vc_do_root_open() to take a vc_device_t* rather than a mx_device_t*.
That was not correct: the value passed is an mx_device_t* but not a
vc_device_t*.

Fortunately, this argument was effectively just used as a boolean
(null vs. non-null).  Change the code to reflect that.

Change-Id: I9cd5bffc62f4c8bcee47ffd202758e47946d257c