[analyze-magenta] Rewrite analyze-magenta script
The current analyze-magenta script no longer work with current magenta build
scripts when using clang toolchain due to the changes in the upstream
clang and magenta build files. This commit replaced the outdated script.
This new script use scan-build-py provided in the latest clang toolchain
to perform static analysis on magenta kernal and generate a report of
defects it found.

Change-Id: I18f530a20923b0154540fde0e6b2f97a8bc1ef53