Track plug state in outputs.
Track the plugged/unplugged state of audio outputs in the AudioServer.
Use it to start implementing a simple set of placeholder routing
policies.  Policies are currently only applied at the time that
AudioRenderers are created.  The next step is to enforce them as
outputs come and go, and their plug state changes.

Change-Id: I70ca780b85de76b79ae45234d72195b5e5747963