apps/media: change demand logic in consumer
This CL changes the way the consumer expresses demand to the upstream
producer. In effect, it establishes a minimum demand of one
outstanding packet. Previously, the demand expressed in the callback
to SupplyPacket was 0 outstanding packets because of the state of the
consumer at that moment. The demand would be updated in a few
milliseconds to 1 packet, requiring a separate round trip. This
effectively doubled the number of round trips required to deliver
a packet across a connection for most connections.

This CL also bumps up the video renderer's demand from 2 to 3
outstanding packets. This was done, because the renderer always keeps
around at least one packet, even if it's stale. This means that the
renderer's need for 2 fresh packets was not getting properly
expressed in starvation scenarios. In the future, the renderer
demand will be expressed in terms of PTS, so the packet retention
won't be an issue. For the moment, this additional demand incurs a
small cost in memory and latency.

Change-Id: I7dd81b4f48828ad9537d82f05b783bfb0b26c36c