apps/media: produce trace events for flog entries
This CL makes flog produce a trace event for each entry logged. The
entry index and channel id are recorded in the trace event. This
makes it easy to correlate flog logs with trace output. Eventually,
I'd like for flog to replace its recording infrastructure with
tracing, but this will be easier once we have fidl 2.0. In the mean
time, these new trace events are extremely helpful.

Change-Id: I4d0f881c1536c56e80d55d4542f48f816ab07caa