[magenta][exceptions] MX_EXCP_THREAD_SUSPENDING -> _SUSPENDED
And similarly MX_EXCP_THREAD_RESUMING -> _RESUMED.
The word tense here was confusing me.
"suspending" is used in the internal context because we're still
in the process of suspending the thread. "suspended" is used in the
external context because once the debugger receives the "suspended"
report it can assume the thread is, for its purposes, now suspended.

Change-Id: If72f5d42db132f9c0fbf0bbd9096533ae9ef0287