Migrate more package configuration to BUILD.gn files
This moves more information about how to assemble packages from JSON
configuration files to BUILD.gn files. In particular, the component
pieces of each package are now expressed with a GN template in
//packages/package.gni.  The template specifies the binaries,
drivers, libraries and resources that map into the assembled archive
along.  The template also has deps on the GN steps required to build
the components of the package.  The JSON configurations contain a
mapping of package names, which must be unique within a build, to the
GN package definitions.  The JSON configurations also contain imports
for other configurations that make sense as a logical connected unit.

The primary advantage of this approach is that more information about
package internals can be written in the same build system rules used
to build the components of the system in an easier to work with
specification language. It is also significantly simpler to locate
information about a package's internal structure along the build
rules and code for the package. This is also a stepping stone towards
having individual packages be more independent instead of relying on
a shared namespace of all files in the system.

To ease rollout, the previous package system is intact in this patch
and works alongside the new system. Once accepted, I can migrate the
rest of the package definitions over to the new system and remove the
implementation code for the old system when the transition is
complete.

Implmentation details:

A new template 'package' is defined. This template contains
information for a single package. At generation time, this produces
manifest files for /system and /boot into
$root_build_dir/$package_name/ and emits ninja rules to assemble the
built artifacts into an archive. It also emits a ninja rule to parse
the build IDs from all compiled artifacts within the package into
$root_build_dir/$package_name/ids.txt.

JSON configuration files now support a new key "packages" that
contains a dictionary of names to labels. The names are package
names, which must match the package_name of the package template
referred to by the label.

packages/gn/gen.py processes the "packages" entries in all imported
configuration files and emits top-level dependencies on all of the
referenced labels as well as an assembly rule that combines the
archives of each package into a top-level archive and the build ids
from each package into a top-level ids.txt.

Once all configurations are migrated to this system, the only
meaningful labels in these configurations will be "packages" and
"imports" to reference additional configurations.

TODO(jamesr): Integrate "components" rules.

Change-Id: Ia7191a9c282b2286925553639c42751dcf759280