[kernel][gigaboot][devmgr] reliable last crashlog mechanism
Introduce platform_stow_crashlog(), a simple API for the panic
path to pass a textual log to the platform in hopes that it will
find its way to the system after a reboot.

For the PC platform, if there is an EFI System Table, attempt to
use EFI Runtime Services to stash this data in an EFI variable.

Gigaboot will read this EFI variable and pass the crashlog to
the booting system as BOOTDATA_LAST_CRASHLOG, erasing the contents
of the variable afterwards.

Devmgr will unpack BOOTDATA_LAST_CRASHLOG into /boot/log/last-panic.txt

Change-Id: I6cc7602a2e85156f8d3353166f4cbc22413a5ae1